+++ Changes.md
@@ -8,3 +8,5 @@
 
 - Refactored middleware support, introducing ability to add client/server middleware (Ryan)
 - Added middleware for ignoring duplicate jobs (Ryan)
+- Added middleware for displaying jobs in resque-web dashboard (Max)
+- Added redis namespacing support (Max)